+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Stuck last replication error, [PR-393](https://github.com/reductstore/reductstore/pull/393)
+
 ## [1.8.0] - 2024-01-24
 
 ### Added